2.3 Asha/Urshu and 2.4 Asha/Susu'Rha
This adds the NPC scenes for Asha Urshu in 2.3 and Asha Susu'Rha in 2.4. I also added the necessary transitions in each file.

For the Asha Urshu scene, since it takes place at the very end of the day, I copied the code for the clock at the end of day and put it in that scene instead, removing it from bargaining debrief.